fix(nl/ics): refactor code. Change names, use one library for testing
[MINI-1658]

+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/nl/creditcards/ICS/filter/ICSRateLimitFilter.java
@@ -7,9 +7,9 @@ public class ICSRateLimitFilter extends RateLimitFilter {
 
     public ICSRateLimitFilter(
             String providerName,
-            long retrySleepMillisecondsMin,
-            long retrySleepMillisecondsMax,
-            long maxRetries) {
+            int retrySleepMillisecondsMin,
+            int retrySleepMillisecondsMax,
+            int maxRetries) {
         super(providerName, retrySleepMillisecondsMin, retrySleepMillisecondsMax, maxRetries);
     }
