+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/nl/creditcards/ICS/filter/ICSRateLimitFilterProperties.java
@@ -5,7 +5,7 @@ import lombok.Value;
 @Value
 public class ICSRateLimitFilterProperties {
 
-    long retrySleepMillisecondsMin;
-    long retrySleepMillisecondsMax;
+    int retrySleepMillisecondsMin;
+    int retrySleepMillisecondsMax;
     int numberOfRetries;
 }
